requirements.txt
google-cloud-storage==2.19.0
google-cloud-core==2.4.1
flask==2.3.2
pymongo==4.3.3
python-dotenv==1.0.0
